Updated "Version Tag Prefix"
Changed version tag prefix from "rel-" to "version-".  We have
an internal concept of a release; e.g. "1713," that is completely
unrelated to the product version number that is produced from
that release.  Yet it is the project version number that would
be important to outside developers and customers when trying to
identify the proper tag in the git repo(s).