updated gitflow init defaults, version 4
Updated the `bin/gitflow-config-init` script as follows:

- To not automatically delete the remote bugfix or feature
  branch when it is finished
- Bump config version number to 4

The reason for the change is that there is a subtle bug with
the gitflow-avh code.  If you have `feature.finish.keepremote`
or `bugfix.finish.keepremote` settings set to false, when you
finish a bugfix or feature it deletes the remote branch--the
copy in origin--before it pushes the updated base branch
back to origin.

This causes pull requests to be closed (which is correct), but
they do not automatically show as being merged (which is not
correct).

If I can get the gitflow-avh folks to fix this, then (when it
eventually propagates to the repos), we can re-enable this
behavior and let gitflow tool clean up the branch in origin for
us.